Use cross-spec section autolinks instead of absolute URLs
This is purely editorial. The spec had a couple of places where it referenced a
section of other specs (RFC9110 and Fetch) using absolute URLs. This update
rather uses the recommended cross-spec section autolink:
https://speced.github.io/bikeshed/#autolink-biblio

This will allow Bikeshed to detect and alert you should the fragment
identifiers change in the targeted specifications, as well as to follow the
specification if its URL changes (the cross-references database typically now
uses the `httpwg.org/specs/` URL for RFC9110 as it is slightly more readable).